+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change log
 
 ## Release 0.1.0
-* Add: Marker Rectangle and Ellipse tool. ([#26](https://github.com/DamirPorobic/kImageAnnotator/issues/26))
+* Add: Double Arrow annotation tool. ([#23](https://github.com/DamirPorobic/kImageAnnotator/issues/23))
+* Add: Marker Rectangle and Ellipse annotation tool. ([#26](https://github.com/DamirPorobic/kImageAnnotator/issues/26))
 * Changed: Blur tool is now preciser and fits the rect. ([#28](https://github.com/DamirPorobic/kImageAnnotator/issues/28))
 * Fixed: Double-click on annotation area causes SIGSEGV crash. ([#29](https://github.com/DamirPorobic/kImageAnnotator/issues/29))